converter/requirements.txt
# Python dependencies for the converter
crewai>=0.165.1
crewai-tools>=0.62.3
pydantic>=2.11.7
PyYAML>=6.0.2

# Development dependencies
pytest==8.4.1
pytest-cov
pytest-mock
